pyproject.toml
[build-system]
requires = ["flit_core>=3.4"]
build-backend = "flit_core.buildapi"

[project]
name = "candl-like"
authors = [
{name = "L. Balkenhol", email = "lennart.balkenhol@iap.fr"},
{name = "C. Trendafilova", email = "ctrendaf@illinois.edu"},
{name = "K. Benabed", email = "benabed@iap.fr"},
{name = "S. Galli", email = "gallis@iap.fr"},
]

readme = "README.rst"
license = {file = "LICENCE.txt"}

dynamic = ["version", "description"]

classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

requires-python = ">=3.9"

dependencies = [
"pyyaml >= 6.0",
"numpy >= 1.23.5",
"scipy >= 1.10.1",
"matplotlib >= 3.7.1",
"clipy-like >= 0.11",
]

[project.optional-dependencies]
docs = [
"sphinx==6.2.1",
"sphinx-collapse==0.1.2",
"sphinx-book-theme==1.0.1",
"numpydoc==1.6.0",
"myst-parser==2.0.0",
"numpy",
]

[tool.flit.module]
name = "candl"

[project.urls]
Source = "https://github.com/Lbalkenhol/candl"
